completed main app styling, started signup page styling

+++ server/data/gamble_create_tables.sql
@@ -36,7 +36,7 @@ CREATE TABLE public.users (
 	"_id" serial NOT NULL,
 	"username" varchar NOT NULL UNIQUE,
 	"password" varchar NOT NULL,
-    "total_cash" bigint NOT NULL,
+    "total_cash" float NOT NULL,
 	CONSTRAINT "users_pk" PRIMARY KEY ("_id")
 ) WITH (
   OIDS=FALSE
